backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.17
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sse-starlette>=1.8.0
python-multipart>=0.0.6
httpx>=0.26.0
langsmith>=0.1.0

